pgbouncer/multi_port/requirements.txt
psycopg2-binary
rich
